Separate ACCRE Slurm correspondence to services.
